move screen & view to presentation
