+++ app/(protected)/index.tsx
@@ -4,7 +4,7 @@ import { ActionButton } from "@/components/actions/ActionButton";
 import { ThemedText } from "@/components/text/ThemedText";
 import { TopView } from "@/feature/dashboard/components";
 import { useLinkInputModal } from "@/feature/links/application/context/LinkInputModalContext";
-import { LinksTopView } from "@/feature/links/views/LinksTopView";
+import { LinksTopView } from "@/feature/links/presentation/views/LinksTopView";
 
 const FloatingButton = () => {
   const { openModal } = useLinkInputModal();
